requirements.txt
streamlit
tensorflow
pillow 
matplotlib